Wind key fix

+++ pl_plot/models.py
@@ -417,7 +417,9 @@ class Overlay(models.Model):
 
         #Delete the Key image
         if os.path.isfile(self.key.path):
-            os.remove(self.key.path)
+            #The wind barb key is static. We don't want to delete it.
+            if self.key.path != 'opt/sharkeyes/media/keys/barbKey.png':
+                os.remove(self.key.path)
 
         directory=os.path.join('/opt/sharkeyes/media/tiles/', self.tile_dir)
